static gen
