+++ package.json
@@ -29,6 +29,7 @@
     "@types/react": "^16.9.43",
     "babel-jest": "^26.1.0",
     "jest": "^26.1.0",
+    "prettier": "^2.5.1",
     "typescript": "^3.9.7"
   }
 }
